Creating the cartSlice:
The createSlice function is called with an object containing configuration options for defining the slice.
The name option specifies the name of the slice, which is used to identify it in the Redux store.
The initialState option specifies the initial state of the slice.
The reducers option is an object where each property represents a Redux action creator function.
In this case, three action creators are defined: addToCart, updateQuantity, and removeFromCart.