Defining the addToCart reducer:
This reducer function is called when an item is added to the cart.
It takes the current state (state) and the action payload (action.payload) as parameters.
It checks if the item being added already exists in the cart by searching for its ID in the cartItems array.
If the item exists, it increments its quantity by 1.
If the item doesn't exist, it adds a new item to the cartItems array with a quantity of 1.