Configuring the Redux store with configureStore:
The configureStore function is called with an object containing configuration options for creating the Redux store.
The reducer option specifies the root reducer for the store. It's an object where each key represents a slice of the Redux state and its corresponding reducer function.
In this case, the cart key is assigned the cartReducer, meaning that the cartReducer handles actions dispatched to the cart slice of the state.
Additional reducers can be added to the reducer object as needed by providing more key-value pairs.
Exporting the Redux store:

The store object created by configureStore is exported as store. This makes the Redux store available for use in other parts of the application.
In summary, this code sets up the Redux store for the application by configuring it with the configureStore function from Redux Toolkit. It specifies the root reducer for the store, which includes the cartReducer to handle actions related to the cart slice of the state. Finally, it exports the configured Redux store for use in other parts of the application.