addToCartHandler function:
This function is called when a user clicks on the "Add To Cart" button for a product.
It checks if the selected product already exists in the cart by comparing its ID with existing cart items.
If the product exists, it dispatches an action to update its quantity by calling the updateQuantity action creator.
If the product doesn't exist, it dispatches an action to add the product to the cart by calling the addToCart action creator.